requirements.txt
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
click==8.1.3
databases==0.7.0
exceptiongroup==1.1.1
fastapi==0.95.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
plotly==5.13.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pytest==7.2.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
